feat: kc node info and detail e2e test
