+++ test/e2e/kcctl/drain_node.go
@@ -1,16 +1,15 @@
-package node
+package kcctl
 
 import (
 	"context"
-	"fmt"
 
 	"github.com/onsi/ginkgo"
 
 	"github.com/kubeclipper/kubeclipper/pkg/query"
 	"github.com/kubeclipper/kubeclipper/pkg/simple/client/kc"
-	"github.com/kubeclipper/kubeclipper/pkg/utils/cmdutil"
 	"github.com/kubeclipper/kubeclipper/test/framework"
 	"github.com/kubeclipper/kubeclipper/test/framework/cluster"
+	fnode "github.com/kubeclipper/kubeclipper/test/framework/node"
 )
 
 var _ = SIGDescribe("[Fast] [Serial] Force drain node", func() {
@@ -43,24 +42,11 @@ var _ = SIGDescribe("[Fast] [Serial] Force drain node", func() {
 
 	ginkgo.It("force drain agent node and ensure the node is deleted", func() {
 		ginkgo.By("drain agent node")
-		err := drainAgentNode(nodeID)
+		err := drainAgentNodeForce(nodeID)
 		framework.ExpectNoError(err)
 
 		ginkgo.By("waiting for node is deleted")
-		err = cluster.WaitForNodeNotFound(f.Client, nodeID, f.Timeouts.CommonTimeout)
+		err = fnode.WaitForNodeNotFound(f.Client, nodeID, f.Timeouts.CommonTimeout)
 		framework.ExpectNoError(err)
 	})
 })
-
-func drainAgentNode(nodeID string) error {
-	return runCmd(fmt.Sprintf("echo y | kcctl drain --agent %s -F", nodeID))
-}
-
-func joinAgentNode(nodeIP string) error {
-	return runCmd(fmt.Sprintf("kcctl join --agent %s", nodeIP))
-}
-
-func runCmd(cmd string) error {
-	ec := cmdutil.NewExecCmd(context.TODO(), "bash", "-c", cmd)
-	return ec.Run()
-}
